+++ TrainingAnalyticsPlatform/storage/table_storage.py
@@ -19,7 +19,7 @@ from azure.storage.blob import BlobServiceClient
 
 from TrainingAnalyticsPlatform.ingestion.fit_parser import compute_workout_id
 
-INGEST_VERSION = "v7.0.0"
+INGEST_VERSION = "v7.0.1"
 CANONICAL_SCHEMA_VERSION = "1.1.0"
 
 WORKOUTS_CONTAINER = "workouts"